refactor: API호출 로직 MainPage로 이동
