+++ client/src/components/FriendsBar/FriendsBar.tsx
@@ -1,43 +1,22 @@
-import { useEffect, useState } from 'react';
-import { FriendsList } from 'GlobalType';
-import { HOST } from '../../constants';
-import axios from 'axios';
+import React from 'react';
+import { Friend } from 'GlobalType';
 import * as S from './FriendsBar.style';
 
-const FriendsBar = () => {
-  const [myProfile, setMyProfile] = useState<FriendsList | null>(null);
-  const [friendsList, setFriendsList] = useState<FriendsList[] | null>(null);
-  const plusIcon = '/plus.svg';
-  const getFriendsList = async () => {
-    try {
-      setFriendsList(null);
-      const response = await axios.get(`${HOST}/api/v1/friend`);
-      setFriendsList(response.data);
-    } catch (error) {
-      console.log(error);
-    }
-  };
-  const getMyProfile = async () => {
-    try {
-      const response = await axios.get(`${HOST}/api/v1/user/me`);
-      setMyProfile(response.data);
-    } catch (error) {
-      console.log(error);
-    }
-  };
+interface FriendsBarProps {
+  myProfile: Friend | null;
+  friendsList: Friend[] | null;
+}
 
-  useEffect(() => {
-    getFriendsList();
-    getMyProfile();
-  }, []);
+const FriendsBar = ({ myProfile, friendsList }: FriendsBarProps) => {
+  const plusIcon = '/plus.svg';
 
   return (
     <>
       <S.FriendsBarContainer>
-        {myProfile && <S.ProfileBox imgURL={myProfile.profileImg}></S.ProfileBox>}
+        {myProfile && <S.ProfileBox imgURL={myProfile.profile_img}></S.ProfileBox>}
         {friendsList &&
-          friendsList.map(({ idx, userId, profileImg }) => {
-            return <S.ProfileBox key={userId} data-idx={idx} imgURL={profileImg}></S.ProfileBox>;
+          friendsList.map(({ idx, userId, profile_img }) => {
+            return <S.ProfileBox key={userId} data-idx={idx} imgURL={profile_img}></S.ProfileBox>;
           })}
         <S.ProfileBox imgURL={plusIcon}></S.ProfileBox>
       </S.FriendsBarContainer>
